[AN-1848] Refactoring Timeline so it has a PostRepository that will serve cards locally(simple cache)/remotelly(ws call) before we treat the memory lea(ws call) before we treat the memory leak..
